web consuming api 2023-0235

+++ SGHT.API/Controllers/TarifasController.cs
@@ -2,6 +2,8 @@
 using SGHT.API.Utils;
 using SGHT.Application.Dtos.Tarifa;
 using SGHT.Application.Interfaces;
+using SGHT.Domain.Base;
+using SGHT.Domain.Entities;
 
 namespace SGHT.API.Controllers
 {
@@ -19,38 +21,75 @@ namespace SGHT.API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetTarifas()
         {
-            var tarifas = await _tarifasService.GetAll();
-            return HandleResponse(tarifas);
+            try
+            {
+                var tarifas = await _tarifasService.GetAll();
+                return HandleResponse(tarifas);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetTarifaById(int id)
         {
-            var tarifa = await _tarifasService.GetById(id);
-            return HandleResponse(tarifa);
+            if (!IsIdValid(id)) return BadRequest("id is not valid");
+            try
+            {
+                var tarifa = await _tarifasService.GetById(id);
+                return HandleResponse(tarifa);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpPost("crear")]
         public async Task<IActionResult> CreateTarifa(SaveTarifaDto tarifas)
         {
-            var result = await _tarifasService.Save(tarifas);
-
-            return HandleResponse(result);
+            if (!IsIdValid(tarifas.IdHabitacion)) return BadRequest("invalid roomid");
+            try
+            {
+                var result = await _tarifasService.Save(tarifas);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpDelete("eliminar")]
         public async Task<IActionResult> EliminarTarifa(DeleteTarifaDto dto)
         {
-            var result = await _tarifasService.DeleteById(dto);
-            return HandleResponse(result);
+            if (!IsIdValid(dto.IdTarifa)) return BadRequest("invalid id");
+            try
+            {
+                var result = await _tarifasService.DeleteById(dto);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpPatch("actualizar")]
         public async Task<IActionResult> ActualizarTarifa(UpdateTarifaDto tarifas)
         {
-            var result = await _tarifasService.UpdateById(tarifas);
-
-            return HandleResponse(result);
+            if (!IsIdValid(tarifas.IdTarifa)) return BadRequest("invalid id");
+            try
+            {
+                var result = await _tarifasService.UpdateById(tarifas);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
     }
 }
